Fix train.py

+++ src/sales_forecast/models/train.py
@@ -43,15 +43,15 @@ def data_preprocessing_pipeline(
     train_df, valid_df, train_df_past, valid_df_past = train_validation_split(data)
 
     scaler_js = JohnsonSU()
-    scaler_js.fit(train_df)
-    train_df = scaler_js.transform(train_df)
-    valid_df = scaler_js.transform(valid_df)
+    scaler_js.fit(train_df.astype(float))
+    train_df = scaler_js.transform(train_df.astype(float))
+    valid_df = scaler_js.transform(valid_df.astype(float))
     X_scaled = pd.concat([train_df, valid_df], axis=0)
     X_scaled = clean_outliers(X_scaled)
 
     scaler_js_past = JohnsonSU()
-    scaler_js_past.fit(train_df_past)
-    X_past_scaled = scaler_js_past.transform(train_df_past)
+    scaler_js_past.fit(train_df_past.astype(float))
+    X_past_scaled = scaler_js_past.transform(train_df_past.astype(float))
 
     clusters = get_clusters(X_scaled)
